WIP: refactor permissions nav structure
