--- ui/business_mng.py
@@ -1,10 +1,0 @@
-import streamlit as st
-from tools.fetch_data import fetch_business
-
-st.header("עסקים")
-
-suppliers = fetch_business()
-if suppliers:
-    st.dataframe(suppliers)
-else:
-    st.info("אין עסקים")
