+++ ui/pages/category_mng.py
@@ -1,0 +1,21 @@
+import streamlit as st
+from tools.api import get, post
+from tools.fetch_data import fetch_categories
+from settings.constants import NAV_CATEGORIES
+
+from tools.helpers import require_permission
+
+
+@require_permission('Admin')
+def category_mng():
+    st.header(NAV_CATEGORIES)
+
+    # List categories
+    cats = fetch_categories()
+    if cats:
+        st.dataframe(cats)
+    else:
+        st.info("אין קטגוריות")
+
+
+category_mng()
