+++ ui/streamlit_app.py
@@ -43,29 +43,6 @@ init_session_state()
 
 st.header(WEBSITE_TITLE)
 st.divider()
-@require_permission('Employee', 'Admin')
-def employee_function():
-    st.title("לעובדים ומנהלים")
-    pages = {
-    NAV_PROJECTS: [
-        st.Page("project_new.py", title=PAGE_NEW, icon=ICON_NEW),
-        st.Page("project_mng.py", title=PAGE_MANAGE, icon=ICON_MANAGE),
-    ],
-    NAV_BUSINESSES: [
-        st.Page("business_new.py", title=PAGE_NEW, icon=ICON_NEW),
-        st.Page("business_mng.py", title=PAGE_MANAGE, icon=ICON_MANAGE),
-    ],
-    NAV_CATEGORIES: [
-        st.Page("category_new.py", title=PAGE_NEW, icon=ICON_NEW),
-        st.Page("category_mng.py", title=PAGE_MANAGE, icon=ICON_MANAGE),
-    ],
-    NAV_OFFERS: [
-        st.Page("offer_new.py", title=PAGE_NEW, icon=ICON_NEW),
-        st.Page("offer_reports.py", title=PAGE_REPORT, icon=ICON_REPORTS),
-    ],
-    }
-    pg = st.navigation(pages)
-    pg.run()
 
 logo_path = Path(WEBSITE_LOGO_PATH)
 
@@ -82,10 +59,16 @@ if not st.session_state.logged_in:
         st.stop()
 else:
     with st.sidebar:
-        st.subheader(f"  👋 יציאה מהמערכת")
         st.write(f"משתמש מחובר  {st.session_state['user']}")
-        logout()
+        if st.button(f"  👋 יציאה מהמערכת"):
+            logout()
 
-employee_function()
+username = get_username()
+print(f"username is : {username}")
+user_permission = get_user_permission_name(username)
+print(f"user_permission is : {user_permission}")
+
+define_sidebar(user_permission)
+
 st.divider()
 # footer
