+++ ui/tools/fetch_data.py
@@ -17,7 +17,6 @@ def fetch_projects() -> list:
     return []
 
 
-
 @st.cache_data(show_spinner=FETCH_CATEGORIES)
 def fetch_categories(project_id: str = None) -> dict:
     url = "/categories"
@@ -60,19 +59,19 @@ def fetch_business():
 
 
 @st.cache_data
-def fetch_business_category(category_id: str = None, business_id:str=None):
+def fetch_business_category(category_id: str = None, business_id: str = None):
     url = "/business-category"
-    
+
     if business_id is None and category_id is None:
         raise ValueError("At least one parameter (category_id or business_id) must be provided")
-        
+
     params = {}
     if category_id is not None:
         params["category_id"] = category_id
-    
+
     if business_id is not None:
         params["business_id"] = business_id
-        
+
     # append query string if params exist
     if params:
         url += f"?{urlencode(params)}"
@@ -121,9 +120,9 @@ def fetch_ai_recom(json_input: dict) -> dict:
         return json.loads(resp.json().get("ai_result"))
     return {}
 
-@st.cache_data
-def fetch_user_details(username:str) -> dict:
-    resp = get("/user/details",json={"username": username})
+@st.cache_data()
+def fetch_user_details(username: str) -> dict:
+    resp = get("/user/details", json={"username": username})
     if getattr(resp, "ok", False):
         return resp.json().get("data", {})
     return {}
@@ -136,6 +135,6 @@ def fetch_permissions() -> list:
     """
     resp = get("/permissions")
     if getattr(resp, "ok", False):
+        print(f"all permissions are: {resp.json()}")
         return resp.json()
     return []
-
